# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AudiobookServer.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/progress": {
      "get": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgressDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/books/{bookId}/progress": {
      "get": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressDto"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/libraries": {
      "get": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnonymousTypeOfGuidAndstringAndstringAndbooleanAndstringAndDateTimeOffsetAndDateTimeOffsetAndint"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/libraries/{id}": {
      "get": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/libraries/{id}/scan": {
      "post": {
        "tags": [
          "AudiobookServer.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "force",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnonymousTypeOfGuidAndstringAndstringAndbooleanAndstringAndDateTimeOffsetAndDateTimeOffsetAndint": {
        "required": [
          "id",
          "name",
          "rootPath",
          "isPublic",
          "credit",
          "lastScanStartedAt",
          "lastScanCompletedAt",
          "books"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "rootPath": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "credit": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastScanStartedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastScanCompletedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "books": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "CreateLibraryRequest": {
        "required": [
          "name",
          "rootPath"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "rootPath": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean",
            "default": false
          },
          "credit": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProgressDto": {
        "required": [
          "bookId",
          "positionSeconds",
          "reportedAt",
          "updatedAt",
          "deviceId",
          "deviceName",
          "isFinished"
        ],
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "positionSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "reportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deviceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "deviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "isFinished": {
            "type": "boolean"
          }
        }
      },
      "ProgressReason": {
        "enum": [
          "first",
          "override",
          "sameDevice",
          "finished",
          "further",
          "behind",
          "behindFinished"
        ]
      },
      "ProgressReport": {
        "required": [
          "bookId",
          "positionSeconds",
          "reportedAt",
          "deviceId",
          "deviceName",
          "isFinished",
          "override"
        ],
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "positionSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "reportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deviceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "deviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "isFinished": {
            "type": "boolean"
          },
          "override": {
            "type": "boolean"
          }
        }
      },
      "ProgressResult": {
        "required": [
          "accepted",
          "reason",
          "progress"
        ],
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "reason": {
            "$ref": "#/components/schemas/ProgressReason"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressDto"
          }
        }
      },
...
      "UpdateLibraryRequest": {
        "required": [
          "name",
          "isPublic",
          "credit"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "credit": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}